requirements.txt
ccxt>=4.2.82
matplotlib>=3.8
numpy>=1.26
optuna>=3.6
pandas>=2.1
PyYAML>=6.0
seaborn>=0.13
pytest>=7.4
